Small refactor of the forms
