+++ src/internal/modules/contact-entry/helpers.js
@@ -1,0 +1,12 @@
+const titleCase = require('title-case');
+const { compact } = require('lodash');
+
+const getAddressText = address => {
+  const { addressLine1, addressLine2, addressLine3, addressLine4 } = address;
+  const addressLines = compact([addressLine1, addressLine2, addressLine3, addressLine4])
+    .map(line => titleCase(line))
+    .join(' ');
+  return `${addressLines}, ${titleCase(address.town)}, ${address.postcode}`;
+};
+
+exports.getAddressText = getAddressText;
